feat: create service for model upload
